refactor(wave0d.1 #252): move shared/ pure-logic to daemon
Move electron/shared/{claudePaths,stateSavedBus}.ts into
packages/daemon/src/shared/. Both modules are pure (zero electron API
imports — only node:path, node:os, node:events). Tests renamed
.test.ts → .spec.ts to match daemon vitest naming. Relative imports get
.js extensions for ESM resolution.

Re-export shims left at original electron/shared/ paths to keep
electron/prefs/{crashReporting,notifyEnabled}.ts compiling — Wave 0d.3
(#249) deletes the shims when prefs/ migrates to the daemon.

[REFACTOR-ONLY] no behavior change.
