+++ packages/daemon/src/shared/__tests__/claudePaths.spec.ts
@@ -10,7 +10,7 @@ import {
   getClaudeConfigDir,
   getClaudeProjectsDir,
   getClaudeSettingsPath,
-} from '../claudePaths';
+} from '../claudePaths.js';
 
 let originalConfigDir: string | undefined;
